pyproject.toml
[project]
name = "pre-commit-ai-checks"
version = "0.1.0"
description = "Pre-commit hooks for checking your projects with AI tools"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pre-commit",
    "python-dotenv",
    "pydantic",
    "pyyaml>=6.0.1",
    "google-genai",
]

[project.scripts]
grammar = "hooks.grammar:main"
silly-mistakes = "hooks.silly_mistakes:main"

[project.optional-dependencies]
dev = ["pytest"]
